+++ bindings/python/iota_sdk/wallet/common.py
@@ -1,11 +1,11 @@
 # Copyright 2023 IOTA Stiftung
 # SPDX-License-Identifier: Apache-2.0
 
-from iota_sdk import call_wallet_method
-import humps
 import json
 from json import dumps, JSONEncoder
 from enum import Enum
+import humps
+from iota_sdk import call_wallet_method
 
 
 def _call_method_routine(func):
@@ -13,18 +13,22 @@ def _call_method_routine(func):
     """
     def wrapper(*args, **kwargs):
         class MyEncoder(JSONEncoder):
-            def default(self, obj):
-                as_dict_method = getattr(obj, "as_dict", None)
-                if callable(as_dict_method):
-                    return obj.as_dict()
-                if isinstance(obj, str):
-                    return obj
-                if isinstance(obj, Enum):
-                    return obj.__dict__
-                if isinstance(obj, dict):
-                    return obj
-                if hasattr(obj, "__dict__"):
-                    obj_dict = obj.__dict__
+            """Custom encoder
+            """
+
+            # pylint: disable=too-many-return-statements
+            def default(self, o):
+                to_dict_method = getattr(o, "to_dict", None)
+                if callable(to_dict_method):
+                    return o.to_dict()
+                if isinstance(o, str):
+                    return o
+                if isinstance(o, Enum):
+                    return o.__dict__
+                if isinstance(o, dict):
+                    return o
+                if hasattr(o, "__dict__"):
+                    obj_dict = o.__dict__
 
                     items_method = getattr(self, "items", None)
                     if callable(items_method):
@@ -32,7 +36,7 @@ def _call_method_routine(func):
                             obj_dict[k] = dumps(v, cls=MyEncoder)
                             return obj_dict
                     return obj_dict
-                return obj
+                return o
         message = func(*args, **kwargs)
 
         for k, v in message.items():
@@ -42,11 +46,10 @@ def _call_method_routine(func):
         def remove_none(obj):
             if isinstance(obj, (list, tuple, set)):
                 return type(obj)(remove_none(x) for x in obj if x is not None)
-            elif isinstance(obj, dict):
+            if isinstance(obj, dict):
                 return type(obj)((remove_none(k), remove_none(v))
                                  for k, v in obj.items() if k is not None and v is not None)
-            else:
-                return obj
+            return obj
         message_null_filtered = remove_none(message)
         message = dumps(humps.camelize(message_null_filtered))
         # Send message to the Rust library
@@ -60,11 +63,9 @@ def _call_method_routine(func):
 
         if "payload" in json_response:
             return json_response['payload']
-        else:
-            return response
+        return response
     return wrapper
 
 
 class WalletError(Exception):
     """A wallet error."""
-    pass